backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart
cors
